add filtering by tags

+++ src-tauri/src/ipc/api.rs
@@ -19,7 +19,7 @@ pub fn remove_tag(app_handle: AppHandle, folder_id: usize, tag_id: usize) -> Ipc
 }
 
 #[command]
-pub fn filter_by_tags(app_handle: AppHandle, tags: Vec<usize>) -> IpcResponse<Vec<String>> {
+pub fn filter_by_tags(app_handle: AppHandle, tags: Vec<u32>) -> IpcResponse<Vec<String>> {
     match app_handle.db(|db| Api::filter_by_tags(db, tags)) {
         Ok(data) => Ok(data).into(),
         Err(e) => Err(Error::Sqlite(e)).into(),
